+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## IN PROGRESS
 
 * Add Jest/RTL testing for `ComponentToPrint` component. Refs UICHKIN-281.
+* Switching service points is not recognized by the check-in app. Refs UICHKIN-344.
 
 ## [7.1.0] (https://github.com/folio-org/ui-checkin/tree/v7.1.0) (2022-06-29)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v7.0.1...v7.1.0)